+++ tests/utils.py
@@ -15,9 +15,11 @@ from __future__ import absolute_import, division, print_function
 
 import collections
 
+import pytest
+
 import six
 
-import pytest
+import cryptography_vectors
 
 
 HashVector = collections.namedtuple("HashVector", ["message", "digest"])
@@ -66,9 +68,8 @@ def check_backend_support(item):
 
 
 def load_vectors_from_file(filename, loader):
-    from cryptography import vectors
-    vector_file = vectors.open_vector_file('hazmat', 'primitives', filename)
-    return loader(vector_file)
+    with cryptography_vectors.open_vector_file(filename) as vector_file:
+        return loader(vector_file)
 
 
 def load_nist_vectors(vector_data):